refactor: consolidate 3 PRs
