+++ src/rules/read-only-properties.ts
